moved trd digitizer to workflow in Detector/TRD/workflow
