+++ Detectors/TRD/CMakeLists.txt
@@ -11,3 +11,4 @@
 add_subdirectory(base)
 add_subdirectory(simulation)
 add_subdirectory(macros)
+add_subdirectory(workflow)
