+++ Detectors/TRD/workflow/include/TRDWorkflow/TRDDigitWriterSpec.h
@@ -27,7 +27,7 @@ namespace trd
 template <typename T>
 using BranchDefinition = framework::MakeRootTreeWriterSpec::BranchDefinition<T>;
 
-o2::framework::DataProcessorSpec getTRDDigitWriterSpec()
+inline o2::framework::DataProcessorSpec getTRDDigitWriterSpec()
 {
   using InputSpec = framework::InputSpec;
   using MakeRootTreeWriterSpec = framework::MakeRootTreeWriterSpec;
