+++ Detectors/TRD/workflow/src/TRDDigitizerSpec.cxx
@@ -8,7 +8,7 @@
 // granted to it by virtue of its status as an Intergovernmental Organization
 // or submit itself to any jurisdiction.
 
-#include "TRDDigitizerSpec.h"
+#include "TRDWorkflow/TRDDigitizerSpec.h"
 #include "Framework/ConfigParamRegistry.h"
 #include "Framework/ControlService.h"
 #include "Framework/DataProcessorSpec.h"
