+++ Steer/DigitizerWorkflow/CMakeLists.txt
@@ -32,7 +32,6 @@ o2_add_executable(digitizer-workflow
                           src/TOFDigitizerSpec.cxx
                           src/TPCDigitRootWriterSpec.cxx
                           src/TPCDigitizerSpec.cxx
-                          src/TRDDigitizerSpec.cxx
                           src/ZDCDigitizerSpec.cxx
                   PUBLIC_LINK_LIBRARIES O2::Framework
                                         O2::Steer
@@ -50,10 +49,11 @@ o2_add_executable(digitizer-workflow
                                         O2::PHOSSimulation
                                         O2::CPVSimulation
                                         O2::TOFSimulation
-					O2::TOFCalibration
+                    					O2::TOFCalibration
                                         O2::TOFReconstruction
                                         O2::TPCSimulation
                                         O2::TPCWorkflow
                                         O2::TRDSimulation
+                                        O2::TRDWorkflow
                                         O2::DataFormatsTRD
                                         O2::ZDCSimulation)
